refs #3 クラスを別ファイルに分割
